+++ packages/sdks/react-sdk/src/components/AuditManagement.tsx
@@ -10,7 +10,7 @@ const AuditManagementWC = lazy(async () => {
   return {
     default: withPropsMapping(
       React.forwardRef<HTMLElement>((props, ref) => (
-	<descope-audit-management-widget ref={ref} {...props} />
+        <descope-audit-management-widget ref={ref} {...props} />
       )),
     ),
   };
@@ -22,11 +22,12 @@ const AuditManagement = React.forwardRef<HTMLElement, AuditManagementProps>(
 
     useImperativeHandle(ref, () => innerRef);
 
-    const { projectId, baseUrl, baseStaticUrl, baseCdnUrl } = React.useContext(Context);
+    const { projectId, baseUrl, baseStaticUrl, baseCdnUrl } =
+      React.useContext(Context);
 
     return (
-	<Suspense fallback={null}>
-		<AuditManagementWC
+      <Suspense fallback={null}>
+        <AuditManagementWC
           projectId={projectId}
           widgetId={widgetId}
           tenant={tenant}
@@ -43,7 +44,7 @@ const AuditManagement = React.forwardRef<HTMLElement, AuditManagementProps>(
             'logger.prop': logger,
           }}
         />
-	</Suspense>
+      </Suspense>
     );
   },
 );