+++ packages/sdks/react-sdk/src/types.ts
@@ -128,6 +128,16 @@ export type DescopeProps = {
   // use to override client context in flow execution
   client?: Record<string, any>;
   styleId?: string;
+  onScreenUpdate?: (
+    screenName: string,
+    context: Record<string, any>,
+    next: (
+      interactionId: string,
+      form: Record<string, any>,
+    ) => Promise<unknown>,
+    ref: HTMLElement,
+  ) => boolean | Promise<boolean>;
+  children?: React.ReactNode;
 };
 
 export type UserManagementProps = WidgetProps;