+++ packages/sdks/web-component/src/lib/types.ts
@@ -10,9 +10,22 @@ export type SdkFlowNext = Sdk['flow']['next'];
 export type ComponentsConfig = Record<string, any>;
 export type CssVars = Record<string, any>;
 
-type OmitFirstArg<F> = F extends (x: any, ...args: infer P) => infer R
-  ? (...args: P) => R
+type KeepArgsByIndex<F, Indices extends readonly number[]> = F extends (
+  ...args: infer A
+) => infer R
+  ? (...args: PickArgsByIndex<A, Indices>) => R
   : never;
+
+type PickArgsByIndex<
+  All extends readonly any[],
+  Indices extends readonly number[],
+> = {
+  [K in keyof Indices]: Indices[K] extends keyof All ? All[Indices[K]] : never;
+};
+
+type Project = {
+  name: string;
+};
 
 export enum Direction {
   backward = 'backward',
@@ -32,8 +45,18 @@ export interface ScreenState {
   form?: Record<string, string>;
   inputs?: Record<string, string>; // Backward compatibility
   lastAuth?: LastAuthState;
+  project?: Project;
   totp?: { image?: string; provisionUrl?: string };
   notp?: { image?: string; redirectUrl?: string };
+  clientScripts?: unknown;
+  selfProvisionDomains?: unknown;
+  user?: unknown;
+  sso?: unknown;
+  dynamicSelects?: unknown;
+  keysInUse?: unknown;
+  genericForm?: unknown;
+  linkId?: unknown;
+  sentTo?: unknown;
 }
 
 export type SSOQueryParams = {
@@ -64,6 +87,7 @@ export type FlowState = {
   baseUrl: string;
   tenant: string;
   stepId: string;
+  stepName: string;
   executionId: string;
   action: string;
   redirectTo: string;
@@ -91,6 +115,8 @@ export type FlowState = {
 
 export type StepState = {
   screenState: ScreenState;
+  screenId: string;
+  stepName: string;
   htmlFilename: string;
   htmlLocaleFilename: string;
   next: NextFn;
@@ -99,11 +125,21 @@ export type StepState = {
   openInNewTabUrl?: string;
 } & OIDCOptions;
 
+export type CustomScreenState = Omit<
+  ScreenState,
+  'cssVars' | 'componentsConfig' | 'inputs'
+> & {
+  error?: {
+    text: ScreenState['errorText'];
+    type: ScreenState['errorType'];
+  };
+};
+
 export type DebugState = {
   isDebug: boolean;
 };
 
-export type NextFn = OmitFirstArg<OmitFirstArg<SdkFlowNext>>;
+export type NextFn = KeepArgsByIndex<SdkFlowNext, [2, 5]>;
 export type NextFnReturnPromiseValue = Awaited<ReturnType<NextFn>>;
 
 export type DebuggerMessage = {
@@ -128,6 +164,7 @@ type Operator =
 
 export interface ClientConditionResult {
   screenId: string;
+  screenName: string;
   interactionId: string;
 }
 
@@ -183,6 +220,7 @@ type ThemeTemplate = {
 
 export type FlowConfig = {
   startScreenId?: string;
+  startScreenName?: string;
   version: number;
   targetLocales?: string[];
   conditions?: ClientCondition[];