+++ packages/libs/sdk-mixins/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.8.0](https://github.com/descope/descope-js/compare/sdk-mixins-0.7.0...sdk-mixins-0.8.0) (2025-02-11)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.2.0`
+* `sdk-component-drivers` updated to version `0.2.51`
+
+### Features
+
+* Custom screens support RELEASE ([#1012](https://github.com/descope/descope-js/issues/1012)) ([20e310d](https://github.com/descope/descope-js/commit/20e310d48f070260a896c9fab0f2b96ef5ccbb3a))
+
 ## [0.7.0](https://github.com/descope/descope-js/compare/sdk-mixins-0.6.7...sdk-mixins-0.7.0) (2025-02-11)
 
 ### Dependency Updates